Move page tree child and parent resolvers to lib/pages
