+++ lib/pages/index.ts
@@ -1,3 +1,4 @@
 export * from './interfaces';
 export * from './getPage';
 export * from './errors';
+export * from './resolvePageTree';
